Bug-Fixing Movement, Grappling Test
Tried out testing a rudimentary grappling hook with the character controller, however I'm not really sure how to make it, good. At the moment it barely pulls the player back and is unable to really modify velocity as we're lacking any real physics simulation. Would like to experiment with converting to a rigidbody.

Minor bug-fixes for movement, move direction works now