Rigidbody Movement, Grappling Hook
Got rid of the player controller and instead opted for a Rigidbody based character controller so that we could have more physics behaviors applying to the player, starting with a grappling hook.

Controller is relatively simple. By creating two checks for if the player is grounded or on a slope (using a sphere-cast at the player's feet every frame) we can check if the player is grounded, by checking the normal direction we can know if they're on a slope. These two control the drag and friction to create smooth feeling and responsive movement, while allowing the player to be affected by other forces.

Created first version of the grappling hook to test this, will both swing and pull the player towards the grapple anchor. The swinging is dependent on your WASD movement rather than your looking. Press F to fire the grapple, release to let go of the grapple.