Redid Level, Started Models
Redid the level, started a large segment with an abyss in the center. Slowly grayboxing the map, creating the shapes of the level. Started on the bottom of the Tower, hanging from the roof. Added a spotlight. Fixed direction bug with firing grappling hook.

Started work on some basic models, one for an arm for the player, the other for the rim of the abyss in the center of the level.