Raycast based Gun, Body Model, UV & GameObjects
Created a very basic hitscan machine gun. It has references to input actions, fill in with the relevant actions for the weapon to fire. Currently just fires raycasts that do nothing, started on the gun prefab, hoping to do recoil next.

Started on a character model, completed the modelling & UV unwrapping, currently working on the texturing.