Body Work, Universal Render Pipeline
Mostly researching today, did a lot of graphics research to try and come up with a technique to achieve Gibbing. Got a couple of different methods I'm going to try.

Placed classes and converted to the Universal Render Pipeline to make use of Decals. Started on rigging up the character model which seems necessary for the Ragdoll effects. Started building out the Body Prefab with Rigidbody components and Character joints which are required for Ragdolling.