Marrying Ragdoll + Character
Working on getting the Ragdoll to play nice with the Character. Current method is to have the Ragdoll as a child of the Character, with the Character bones moving the ragdoll bones to match. When the character dies, it releases and enables the ragdoll and deletes itself.

BodyPart performs the separation; when you kill an enemy the last limb damaged will separate the corresponding bone & renderer on the ragdoll. Originally this was munted as each limb is affected by multiple bones; by removing the multi-limb weighting we get blocky looking movement however, it separates nicely.

Need to redo configuration methods, currently the hierarchy is being turned into a mess because of this. We're just automatically combing through ragdoll objects to find bones that match the body, however these aren't being handled as I expected and some bones (such as shoulder bones) complicate this further. Want to switch to manual configurations via the editor; we can only have one humanoid shape so manual work is acceptable.