Ragdoll & Limb Detach on Death
Finished marrying up the ragdoll to the avatar, ragdoll will now match the position and be correctly detached when destroying a limb. Created extremely early gibs for the Hip Bone, experimenting by just adding a random force to each Gib on death. There is an issue with the torso where one of the legs is disappearing as a ragdoll. Not sure what could be triggering that yet.

Added Hands & Feet to the ragdoll to support how bodyparts have been configured. Next up is animations & particle effects.