Re-Rigging, Model updates, Test VFX
Did testing on Gibs & particle effects, tried to create some test blood particles with vague textures. Created a line renderer for the grappling hook to show the rope. Edited the character model to include an Abdomen that can now be detached & destroyed. Character prefab was unlinked causing issues when this happened, rebuilt the prefabs to properly link back to model changes. Ragdoll & Body have both been set up as their own prefabs.

Did a lot of changes to how Bodies & Bodyparts detach and moved the data to be built on the Bodypart layer, making it simpler to configure. Created running & idle animations.