Blood Particles & Gibs
Created Gibs for each limb with particle systems configured to create blood and leave blood particles on the floor. Gibs explode out and fly everywhere, destroyed joints have their own gibs attached to them. Gibs are created on the Ragdoll, simply place a physics object on the ragdoll as a child with the Giblet tag and the system will use it as a Gib.

Currently a bug where the right thigh is being consistently destroyed with the abdomen.