Crouching First Attempt
Attempted to use Additive animations for getting crouching working, instead caused issues with the orientation of the model and had every limb key-framed each frame. Instead thinking, if I can limit the keyframes to the specific limbs moving in each animation, in theory the additive animation should work.

My idea is that if I restrict the axis on each keyframe in each animation, theoretically I won't have to make separate animations for crouching, pending on what the result looks like. I think this would be a really neat way to handle crouching so I'd like to attempt again.