Player Upgrades & Legs
Separated all of the different functions of the Player into different Scripts allowing us to put more depth into each:

Improved crouching so that it checks if the Player has an obstacle above their head and prevents standing back up if the Player is unable to. Crouching has better Camera & Ground check controls with inspector variables to tweak.

Grappling behaves very differently. Instead of modifying the velocity we're just pulling the player towards the anchor, added some threshold levels and a progress check that modify movement speed, grapple speed and gravity the closer you get to it, feels more responsive now.

Made Movement Modifier a public property. This will change when OOP stats are implemented, for now just remember what the modifier was before you change it, then set it back when you're done.

Added Legs to the Player, need to redo the prefab so we can preserve the upper body bones we won't see. This is so that we can still create the ragdoll on Player death in the matching pose.

Moved Camera position change to Update instead of Fixed Update, fixed the desync between camera position and model so you no longer run ahead of your own camera.

Added some more keyframes to the crouch pose, was working on proper animations before I got blue screened :(