Buckshot Pistol & Testing new Workflow
Trying out a different method of creating new animations in a separate file to workaround Unity not importing all animations in the original file. Haven't been able to test that stage yet. Need to change the modifier in the Player Arms model so that the hands rotate independently of arm orientation, help us get more control over it.

Created a simple Buckshot Pistol, with the idea that its two barrels and each fires a single round of buckshot, needing to be reloaded after just two shots.