Grapple Animations & Arm Changes
Changed the relationship of the hand bones to the arms, didn't have a lot of time today to do much so simply updated that and turned the Grapple animation into two animations:
First one is when firing the grapple, arm extends forward, balls into a fist and recoils backwards when its meant to fire.
Second animation is when it hits, arm rotates around and sharply pulls back towards the player's head, recoiling a little.