Level Work
Expanded on the Net Labs part of the level and tested the navigation with a simple script to follow a target which works.

Net Labs now has a cooling tower accessible by two different vent systems, created some simple geometry for aesthetic pieces to make it look cool and restrict player movement, trying to create a claustrophobic spiral in the tower to connect it to the large elevator shaft.