Post FX & More Blocking
Working on practising the level geometry a lot, working the doorways into the scene better so that they'll play nice with the Navmesh which now looks quite nice in the Netlabs. Expanded more of the Coolant tower to restrict access between the two main paths the player can take and adding to the maintenance catwalk surrounding it.

Mucked around with Post FX to make darks darker, think it looks kinda nifty but need to see more of the materials in place before I can know if it works well, at least the bloom looks nice on the player arms which I'm really happy with.